Fix: bug runtime Vanilla - wx.Rect/DrawText/Gauge refusent les float
BUG RUNTIME BLOQUANT reproduit reellement en reconstruisant et en
lancant l'executable Windows portable produit par la CI (le job
"Construire la Vanilla Windows" etait vert car son smoke test se
contente de verifier que le processus reste actif, sans verifier
qu'une fenetre principale apparait reellement) :

    OnInit -> OuvrirDernierFichier -> OuvrirFichier ->
    AfficherServeurConnecthys -> CTRL_Portail_serveur.MAJ_bouton ->
    CTRL_TaskBarIcon.Connecthys -> AjouteTexteImage
    TypeError: Rect(): ... argument has unexpected type 'float'

Cause : "hauteurRond / 2.0" (division reelle) produit toujours un
float, transmis tel quel a wx.Rect(...) puis a dc.DrawText(...) ;
wxPython Phoenix rejette les arguments float la ou wx Classic les
acceptait silencieusement. Confirme independant de la PR Convention
(aucun commit de cette PR ne touche ces fichiers ; le meme code est
present sur maintenance/vanilla).

Ratissage de la meme famille (fonction "rond rouge avec texte"
dupliquee dans 5 fichiers, wx.Rect et wx.Gauge) :
- CTRL_TaskBarIcon.py, DLG_Envoi_sms.py, DLG_Ouvertures.py,
  DLG_Selection_mails.py : meme defaut confirme sur wx.Rect ET sur
  dc.DrawText (corriger seulement le Rect aurait juste deplace le
  crash un peu plus loin) ;
- CTRL_Commande_repas.py : deja correctement caste en int, faux
  positif ecarte (verifie avant toute correction) ;
- UTILS_Portail_synchro.Pulse_gauge : meme famille, revelee apres la
  premiere correction (le crash precedent empechait d'atteindre ce
  code) -- "100 * num_etape / nbre_etapes" transmis tel quel a
  wx.Gauge.SetValue via wx.CallAfter, corrige de la meme facon.

Reconstruit reellement l'executable en local (pyinstaller, meme spec
que la CI) et relance : fenetre principale visible, stderr propre,
plus de trace.

Ajoute tests/test_vanilla_rect_float_regression.py (7 tests, execute
les memes fonctions avec les memes types d'arguments que les appels
reels qui ont crashe) et le fait tourner dans le workflow Windows
("Construire la Vanilla Windows") pour exercer le vrai runtime wx
avant meme de construire l'executable.

N'ajoute ni try/except ni masquage : les valeurs sont normalisees en
entier a la source, le rendu visuel est conserve, Connecthys() n'est
pas modifie fonctionnellement.

PR #370 reste en Draft. Aucun developpement Convention dans ce commit.

+++ noethys/Utils/UTILS_Portail_synchro.py
@@ -104,7 +104,7 @@ class Synchro():
             self.num_etape += 1
         else :
             self.num_etape = num
-        num = 100 * self.num_etape / self.nbre_etapes
+        num = int(100 * self.num_etape / self.nbre_etapes)
         try :
             self.log.SetGauge(num)
         except :
